lib/udelay.c: avoid a division in udelay()
The sipler original code forced a division, at least with some compiler,
even if the dividend is constant.  This resulted in misworking w1
if this delay was used on lpc2104.

The commit replaces the division with a multiplication, which allows
removal of arch-specific udelay support in lpc2104, and use this one
instead.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>