build: link bathos with $(LD), but arch may override
The makefile used $(CC) to link bathos.o into bathos. This worked
as long as the cross-compiler is a no-libc build, but if a full compiler
with default libc is used, it picked its own library.

One solution is using "-nostdlib" and another is linking with $(LD)
instead.  Both options spit a warning on native builds, where the
default library is used.

This commit thus introduced $(ARCH-FINAL-LD) which is $(LD)
by default, but arch-unix sets it to $(CC).

Reported-by: Filippo Scognamiglio <flscogna@gmail.com>
Signed-off-by: Alessandro Rubini <rubini@gnudd.com>